+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 ### New Features
 - Add in launcher config to be able to toggle and disable things without a launcher update
 - Add initial Quilt loader support (disabled)
+- Add warning prompt when adding a mod with type Inside Minecraft.jar
 
 ### Fixes
 - Update the LaunchServer.bat script to make it "simpler" to change Java path